Move functions to occlient
